extract Player class to separate file
