+++ models.py
@@ -6,42 +6,6 @@ import pygame as pg
 from settings import Settings
 
 
-class Player(pg.sprite.Sprite):
-    """Representing the player as a moon buggy type car."""
-
-    speed = 10
-    bounce = 24
-    gun_offset = -11
-    images: List[pg.Surface] = []
-
-    def __init__(self, *groups):
-
-        self.settings = Settings()
-        self.SCREENRECT = self.settings.SCREENRECT
-
-        pg.sprite.Sprite.__init__(self, *groups)
-        self.image = self.images[0]
-        self.rect = self.image.get_rect(midbottom=self.SCREENRECT.midbottom)
-        self.reloading = 0
-        self.origtop = self.rect.top
-        self.facing = -1
-
-    def move(self, direction):
-        if direction:
-            self.facing = direction
-        self.rect.move_ip(direction * self.speed, 0)
-        self.rect = self.rect.clamp(self.SCREENRECT)
-        if direction < 0:
-            self.image = self.images[0]
-        elif direction > 0:
-            self.image = self.images[1]
-        self.rect.top = self.origtop - (self.rect.left // self.bounce % 2)
-
-    def gunpos(self):
-        pos = self.facing * self.gun_offset + self.rect.centerx
-        return pos, self.rect.top
-
-
 class Alien(pg.sprite.Sprite):
     """An alien space ship. That slowly moves down the screen."""
